Add rating feature

+++ templates/views/view.php
@@ -16,29 +16,33 @@
     <div class="top-bar top-bar--column">
       <div class="page-title h1"><?= $recipe->getName() ?></div>
       <div class="recipe-actions">
-        <?php createStars($recipe->getStars()) ?>
-        <div class="button button--menu">
-          <i class="fa-solid fa-ellipsis"></i>
-          <div class="button-menu">
-            <?php
-            if ($canEdit) { ?>
-              <a class="button" href="/edit?id=<?= $recipe->getId() ?>"><i class="fa-solid fa-pen"></i> Edit</a>
-            <?php } ?>
-            <?php
-            if ($canPublish) { ?>
-              <a class="button" href="/publish?id=<?= $recipe->getId() ?>"><i class="fa-solid fa-eye"></i> Publish</a>
-            <?php } ?>
-            <?php
-            if ($canUnpublish) { ?>
-              <a class="button" href="/unpublish?id=<?= $recipe->getId() ?>"><i class="fa-solid fa-eye-slash"></i>
-                Unpublish</a>
-            <?php } ?>
-            <?php
-            if ($canDelete) { ?>
-              <a class="button" href="/delete?id=<?= $recipe->getId() ?>"><i class="fa-solid fa-xmark"></i> Delete</a>
-            <?php } ?>
-          </div>
+        <div class="stars-rate">
+          <?php createStars($recipe->getStars()) ?>
         </div>
+        <?php if ($canEdit || $canPublish || $canUnpublish || $canDelete) { ?>
+          <div class="button button--menu">
+            <i class="fa-solid fa-ellipsis"></i>
+            <div class="button-menu">
+              <?php
+              if ($canEdit) { ?>
+                <a class="button" href="/edit?id=<?= $recipe->getId() ?>"><i class="fa-solid fa-pen"></i> Edit</a>
+              <?php } ?>
+              <?php
+              if ($canPublish) { ?>
+                <a class="button" href="/publish?id=<?= $recipe->getId() ?>"><i class="fa-solid fa-eye"></i> Publish</a>
+              <?php } ?>
+              <?php
+              if ($canUnpublish) { ?>
+                <a class="button" href="/unpublish?id=<?= $recipe->getId() ?>"><i class="fa-solid fa-eye-slash"></i>
+                  Unpublish</a>
+              <?php } ?>
+              <?php
+              if ($canDelete) { ?>
+                <a class="button" href="/delete?id=<?= $recipe->getId() ?>"><i class="fa-solid fa-xmark"></i> Delete</a>
+              <?php } ?>
+            </div>
+          </div>
+        <?php } ?>
       </div>
     </div>
     <div class="recipe-details">
@@ -51,9 +55,18 @@
         <div class="role"><?= $author->getRoleName() ?></div>
       </div>
       <div class="recipe-details__recipe">
-        <div id="recipe-content" class="recipe-details__content"
+        <div id="recipe-content" class="recipe-details__content markdown-body"
              data-md="<?= htmlspecialchars($recipe->getDetails() ?? '') ?>"></div>
       </div>
+      <?php if ($canRate) { ?>
+        <hr>
+        <div class="rate-section">
+          <span>Did this recipe delight your taste buds?</span>
+          <div id="stars-rate">
+            <?php createStars($userRating) ?>
+          </div>
+        </div>
+      <?php } ?>
     </div>
   </main>
 </div>
